Move fixtures to separate module
